[code] avoid unrequired unboxing

+++ code/src/java/pcgen/io/exporttoken/WeaponToken.java
@@ -878,7 +878,7 @@ public class WeaponToken extends Token
 	public static String getRangeToken(Equipment eq, PlayerCharacter pc, boolean units)
 	{
 		StringBuilder sb = new StringBuilder();
-		sb.append(Globals.getGameModeUnitSet().displayDistanceInUnitSet(EqToken.getRange(pc, eq).intValue()));
+		sb.append(Globals.getGameModeUnitSet().displayDistanceInUnitSet(EqToken.getRange(pc, eq)));
 
 		if (units)
 		{
@@ -2414,7 +2414,7 @@ public class WeaponToken extends Token
 			{
 				if ((damString.charAt(index) == '+') || (damString.charAt(index) == '-'))
 				{
-					totalBonus = Delta.decode(damString.substring(index)).intValue();
+					totalBonus = Delta.decode(damString.substring(index));
 					break;
 				}
 			}
@@ -2960,7 +2960,7 @@ public class WeaponToken extends Token
 	public static List<String> getRangeList(Equipment eq, boolean addShortRange, final PlayerCharacter aPC)
 	{
 		final List<String> aList = new ArrayList<>();
-		final int baseRange = EqToken.getRange(aPC, eq).intValue();
+		final int baseRange = EqToken.getRange(aPC, eq);
 		int aRange = baseRange;
 		int maxIncrements = 0;
